Migrate usage of metrics.Handler to metricDefinition.With
Using the new metricDefinition.With approach.

The usage of metrics.Handler could be both redundant and unsafe, so
using the new With(Handler) method, there's no need to specify redundant information.

Already exists an ExampleHandler which tests all the new approach methods and also serves as documentation on how to use the Handler, to encourage people to use the new approach.

People can still use the old approach because we are keeping the Handler interface as-is.

No.